refactor: split ui.Select type
