+++ disnake/ui/__init__.py
@@ -14,6 +14,7 @@ from .action_row import *
 from .button import *
 from .item import *
 from .modal import *
-from .select import *
+from .select_base import *
+from .select_string import *
 from .text_input import *
 from .view import *
